clarify relationship between substring functions
The three functions STRPOS, CHARINDEX, and POSITION
are for the same purpose (to get the index of a substring
in a string), but they are described as "synonyms"
currently in the docs, which they are not, as they do not
have identical meaning (the argument structures for all
three of them are different. The word "synonym" is confusing
because it implies that they are aliases of each other.

So here we say "similar", and also we take off the note
at the bottom of one of them saying to look to the other
one for "details". They all have pretty much the same
information, so that's not really necessary or helpful. And
their relationship is flagged at the top of the article anyway.